+++ package.json
@@ -35,6 +35,8 @@
     "@mui/material": "^5.8.0",
     "@mui/styled-engine-sc": "^5.7.0",
     "@mui/styles": "^5.8.0",
+    "@mui/x-data-grid": "^5.13.0",
+    "@mui/x-data-grid-generator": "^5.13.0",
     "axios": "^0.27.2",
     "file-loader": "^6.2.0",
     "hangul-romanization": "^1.0.0",
